Id as a separate class record.
