+++ backend/src/PetFamily.Domain/Volunteers/PaymentDetails.cs
@@ -1,4 +1,4 @@
-namespace PetFamily.Domain.Volunteer;
+namespace PetFamily.Domain.Volunteers;
 
 public class PaymentDetails
 {
