+++ backend/src/PetFamily.Domain/Volunteers/PetId.cs
@@ -1,0 +1,13 @@
+namespace PetFamily.Domain.Volunteers;
+
+public record PetId
+{
+    private PetId(Guid value)
+    {
+        Value = value;
+    }
+    public Guid Value { get; }
+    
+    public static PetId NewPetId() => new(Guid.NewGuid());
+    public static PetId Empty() => new(Guid.Empty);
+}
