+++ backend/src/PetFamily.Domain/Volunteers/VolunteerId.cs
@@ -1,0 +1,14 @@
+namespace PetFamily.Domain.Volunteers;
+
+public record VolunteerId
+{
+    private VolunteerId(Guid value)
+    {
+        Value = value;
+    }
+    
+    public Guid Value { get; }
+    
+    public static VolunteerId NewVolunteerId => new(Guid.NewGuid());
+    public static VolunteerId Empty => new(Guid.Empty);
+}
